Create README.md
EDUDORM is a freelance writing company that offers the best, high quality, and non-plagiarized custom papers. We are committed to providing only quality papers that meet our clients’ expectations. Having been in the writing industry for a couple of years now, we have an understanding of what every customer wants and we always meet those expectations. We provide all writing services such as writing from scratch, proofreading, re-writing, and editing. All our writers are truly hard workers and they are always committed to providing the best services ever. We provide services 24/7, meaning we are always there to assist our esteemed clients. We always put the interests of our clients first because we know it is through the excellent experience that customers get from us that makes them come back.
https://www.edudorm.com/